custom invoices to statistics, first tests, refactor tests a bit
